moving RHS function calling into solver
