+++ qiskit/providers/aer/pulse/de_solvers/pulse_de_options.py
